Enable 256-bit AVX and improve simd::gather intrinsics
Align data to 32 bytes for fast AVX-256 loads. Use vector loads for
gather indices for improved performance with all instruction sets.
Use AVX2 gather intrinsics when possible -- results in a nice ~10%
boost for ELLPACK matrix-vector multiplication with 32-bit floats.

+++ cppcore/include/kpm/OptimizedHamiltonian.hpp
@@ -124,7 +124,7 @@ public:
     template<class Vector>
     Vector reorder_vector(Vector const& v) const {
         if (reorder_map.empty()) { return v; }
-        assert(reorder_map.size() == v.size());
+        assert(reorder_map.size() == static_cast<size_t>(v.size()));
 
         auto result = Vector(v.size());
         for (auto i = 0; i < result.size(); ++i) {
